Retry Kafka topic description while the broker does not know the topic yet
Right after a topic is created, the single KRaft broker's metadata can
trail the controller, so describeTopics fails with
UnknownTopicOrPartitionException for a topic listTopics just returned.
Topic provisioning then threw "Failed to provision Kafka topics" instead
of provisioning or reporting the partition mismatch (Kafka CI runs
34346446216 and 32527275388, KafkaReusedContainerTopicProvisioningTest).

retryMetadataRequest now also retries that failure after a 1s delay and
rethrows it once the attempts are exhausted. Other execution failures are
still rethrown immediately.

The broker-free admin retry specs move to their own KafkaAdminRetryTest,
and the aggregate build runs that spec without -Pkafka instead of
disabling the module's tests, so JaCoCo and Sonar see the retry code
without starting any Kafka container. -Pkafka still runs everything.
